Task: update to TYPO3 v8
To make the extension installable in TYPO3 v8 instance the emconf needed to be updated.
Since v8 relys on the 4.0 branch of the symfony css-selector package this is a breaking change.

Another BC is the way how the database connection settings are stored in the additional configuration and how to optain them.